% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLon2Basin.R
\name{LatLon2Basin}
\alias{LatLon2Basin}
\title{Thai River basin name}
\usage{
LatLon2Basin(lat, lon)
}
\arguments{
\item{lat}{numeric, latitude}

\item{lon}{numeric, longitude}
}
\value{
Main River basin name
}
\description{
Finding the main River basin name using Latitude and longitude
}
\examples{
LatLon2Basin(13.7506,100.7943)
LatLon2Basin(14.074251, 100.608416)
LatLon2Basin(18.7883, 98.9853)

}
