% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLon2BasinTbl.R
\name{LatLon2BasinTbl}
\alias{LatLon2BasinTbl}
\title{Thai River basin name}
\usage{
LatLon2BasinTbl(LatLon)
}
\arguments{
\item{LatLon}{numeric, lat lon in data frame}
}
\value{
Main River basin name
}
\description{
Finding the main River basin name using Latitude and longitude
}
\examples{
LatLon2BasinTbl(LatLon)

}
